% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_data}
\alias{build_data}
\title{Title}
\usage{
build_data(
  authors,
  filiation,
  title,
  abstract,
  keywords,
  references,
  pages,
  year,
  volume,
  number,
  language,
  doi,
  x,
  pdf_url,
  full_text,
  journal,
  issn
)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
