% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartainternacional.R
\name{cartainternacional}
\alias{cartainternacional}
\title{Title}
\usage{
cartainternacional(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
