% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunturaglobal.R
\name{conjunturaglobal}
\alias{conjunturaglobal}
\title{Title}
\usage{
conjunturaglobal(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
