% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunturainternacional.R
\name{conjunturainternacional}
\alias{conjunturainternacional}
\title{Title}
\usage{
conjunturainternacional(
  year,
  volume,
  number,
  silence = TRUE,
  full_text = FALSE
)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
