% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunturaaustral.R
\name{conjunturalaustral}
\alias{conjunturalaustral}
\title{Title}
\usage{
conjunturalaustral(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
