% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estudosinternacionais.R
\name{estudosinternacionais}
\alias{estudosinternacionais}
\title{Title}
\usage{
estudosinternacionais(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
