% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luanova.R
\name{luanova}
\alias{luanova}
\title{Title}
\usage{
luanova(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{volume}{Note: When volume == "s/v" it should be set as 0 in the function call}

\item{full_text}{}
}
\value{

}
\description{
Title
}
