% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meridiano47.R
\name{meridiano47}
\alias{meridiano47}
\title{Title}
\usage{
meridiano47(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{number}{Note: From 2016 editions have only volume and the number should be set to 0}

\item{full_text}{}
}
\value{

}
\description{
Title
}
