% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muralinternacional.R
\name{muralinternacional}
\alias{muralinternacional}
\title{Title}
\usage{
muralinternacional(year, volume, number, silence = TRUE, full_text = FALSE)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
